Moving functions to extra file for better maintenance and overview
